Addressing Matt's comments (part3): Moving _get_task_dict test to integration testing
