+++ tests/unit/evaluation/test_lm_harness.py
@@ -2,7 +2,6 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 import torch
-from lm_eval.api.group import ConfigurableGroup
 from lm_eval.api.task import ConfigurableTask
 
 from oumi.core.configs import (
@@ -15,7 +14,6 @@ from oumi.core.configs import (
 from oumi.core.configs.params.evaluation_params import EvaluationPlatform
 from oumi.evaluation.lm_harness import (
     _generate_lm_harness_model_args,
-    _get_task_dict,
     evaluate,
 )
 
@@ -224,55 +222,6 @@ def test_generate_lm_harness_model_args(
         mock_build_processor.assert_not_called()
 
     assert model_args == expected_model_args
-
-
-def test_get_task_dict_for_configurable_task():
-    task_params = LMHarnessTaskParams(
-        evaluation_platform="lm_harness",
-        task_name="mmlu_college_computer_science",
-        num_fewshot=33,
-    )
-
-    task_dict = _get_task_dict(task_params)
-
-    assert len(task_dict) == 1
-    assert "mmlu_college_computer_science" in task_dict
-    task: ConfigurableTask = task_dict["mmlu_college_computer_science"]  # type: ignore
-
-    assert task.config.task == "mmlu_college_computer_science"
-    assert task.config.num_fewshot == 33
-    assert len(task.eval_docs) == 100
-    assert task.OUTPUT_TYPE == "multiple_choice"
-
-
-@pytest.mark.skip(reason="Temporarily disabled because it times out.")
-def test_get_task_dict_for_configurable_group():
-    task_params = LMHarnessTaskParams(
-        evaluation_platform="lm_harness", task_name="mmmu_val", num_fewshot=222
-    )
-
-    task_dict = _get_task_dict(task_params)
-
-    # Top Level: A single ConfigurableGroup with 6 subgroups
-    assert len(task_dict) == 1
-    conf_group_key = next(iter(task_dict))
-    assert isinstance(conf_group_key, ConfigurableGroup)
-    assert conf_group_key.group == "mmmu_val"
-    conf_group_dict = task_dict[conf_group_key]
-    assert isinstance(conf_group_dict, dict)
-    assert len(conf_group_dict) == 6
-
-    # Subgroup level: ConfigurableGroups consisting of multiple tasks
-    for subgroup_key, subgroup_dict in conf_group_dict.items():
-        assert isinstance(subgroup_key, ConfigurableGroup)
-        assert isinstance(subgroup_dict, dict)
-
-        # Task level: ensure `num_fewshot` has propagated to all tasks.
-        for task_key, task in subgroup_dict.items():
-            assert isinstance(task_key, str)
-            assert task_key.startswith("mmmu_val")
-            assert isinstance(task, ConfigurableTask)
-            assert task.config.num_fewshot == 222
 
 
 def test_evaluate(mock_patches_for_evaluate):
